o Applied patch in bug [1006218] getObject returns incorrect Java class;    (thanks to Mike Hutchinson for the patch).
git-svn-id: http://svn.code.sf.net/p/jtds/code/trunk/jtds@503 97a8069c-bbaa-4fa2-9e49-561d8e0b19ef

+++ src/main/net/sourceforge/jtds/jdbc/JtdsCallableStatement.java
@@ -47,7 +47,7 @@ import java.util.TimeZone;
  * </ol>
  *
  * @author Mike Hutchinson
- * @version $Id: JtdsCallableStatement.java,v 1.3 2004-08-05 01:45:22 ddkilzer Exp $
+ * @version $Id: JtdsCallableStatement.java,v 1.4 2004-08-10 14:57:44 bheineman Exp $
  */
 public class JtdsCallableStatement extends JtdsPreparedStatement implements CallableStatement {
     /**
@@ -115,7 +115,7 @@ public class JtdsCallableStatement extends JtdsPreparedStatement implements Call
     public byte getByte(int parameterIndex) throws SQLException {
         ParamInfo pi = getParameter(parameterIndex);
 
-        return ((Byte) Support.convert(this, pi.value, java.sql.Types.TINYINT, null)).byteValue();
+        return ((Integer) Support.convert(this, pi.value, java.sql.Types.TINYINT, null)).byteValue();
     }
 
     public double getDouble(int parameterIndex) throws SQLException {
@@ -127,7 +127,7 @@ public class JtdsCallableStatement extends JtdsPreparedStatement implements Call
     public float getFloat(int parameterIndex) throws SQLException {
         ParamInfo pi = getParameter(parameterIndex);
 
-        return ((Float) Support.convert(this, pi.value, java.sql.Types.FLOAT, null)).floatValue();
+        return ((Double) Support.convert(this, pi.value, java.sql.Types.FLOAT, null)).floatValue();
     }
 
     public int getInt(int parameterIndex) throws SQLException {
@@ -145,7 +145,7 @@ public class JtdsCallableStatement extends JtdsPreparedStatement implements Call
     public short getShort(int parameterIndex) throws SQLException {
         ParamInfo pi = getParameter(parameterIndex);
 
-        return ((Short) Support.convert(this, pi.value, java.sql.Types.SMALLINT, null)).shortValue();
+        return ((Integer) Support.convert(this, pi.value, java.sql.Types.SMALLINT, null)).shortValue();
     }
 
     public boolean getBoolean(int parameterIndex) throws SQLException {
